requirements.txt
APScheduler==3.11.0
discord.py==2.4.0
langchain==0.3.14
langchain_openai==0.3.1
python-dotenv==1.0.1
